+++ src/types/creep/transporter.d.ts
@@ -1,0 +1,34 @@
+// 运输相关方法接口扩展
+interface Creep {
+  /**
+   * 检查房间的孵化能量是否足够，
+   * 不足的情况下会设工作状态为WORK_TRANSPORTER_SPAWN
+   * @returns 是否成功变化工作状态
+   */
+  checkWorkTransporterSpawn(): boolean;
+  /**
+   * 执行WORK_TRANSPORTER_SPAWN
+   */
+  doWorkTransporterSpawn(): void;
+
+  /**
+   * 检查房间的塔能量是否需要补充
+   * 有需要补充能量的塔时，设工作状态为WORK_TRANSPORTER_TOWER
+   * @returns 是否成功变化工作状态
+   */
+   checkWorkTransporterTower(): boolean;
+   /**
+    * 执行WORK_TRANSPORTER_TOWER
+    */
+   doWorkTransporterTower(): void;
+  /**
+   * 检查房间的source container是否接近满了
+   * 如果有，设工作状态为WORK_TRANSPORTER_STORAGE
+   * @returns 是否成功变化工作状态
+   */
+   checkWorkTransporterStorage(): boolean;
+   /**
+    * 执行WORK_TRANSPORTER_TOWER
+    */
+    doWorkTransporterStorage(): void;
+}
